# audit: avoid extra byte slice to string convertion