<!--  Thanks for sending a pull request!  Here are some tips for you:

1. If this is your first time, please read our contributor guidelines: https://git.k8s.io/community/contributors/guide/first-contribution.md#your-first-contribution and developer guide https://git.k8s.io/community/contributors/devel/development.md#development-guide
2. Please label this pull request according to what type of issue you are addressing, especially if this is a release targeted pull request. For reference on required PR/issue labels, read here:
https://git.k8s.io/community/contributors/devel/sig-release/release.md#issuepr-kind-label
3. Ensure you have added or ran the appropriate tests for your PR: https://git.k8s.io/community/contributors/devel/sig-testing/testing.md
4. If you want *faster* PR reviews, read how: https://git.k8s.io/community/contributors/guide/pull-requests.md#best-practices-for-faster-reviews
5. If the PR is unfinished, see how to mark it: https://git.k8s.io/community/contributors/guide/pull-requests.md#marking-unfinished-pull-requests
-->

#### What type of PR is this?
/kind bug
<!--
Add one of the following kinds:
/kind bug
/kind cleanup
/kind documentation
/kind feature

Optionally add one or more of the following kinds if applicable:
/kind api-change
/kind deprecation
/kind failing-test
/kind flake
/kind regression
-->

#### What this PR does / why we need it:
Remove extra byteslice to string conversion when writing audit log. Benchtest results for `go test -benchtime 1s -count 6 -cpu 1 -benchmem -timeout 999m --bench=BenchmarkLogEvents -run '^$'`:
```
goos: linux
goarch: amd64
pkg: k8s.io/apiserver/plugin/pkg/audit/log
cpu: AMD Ryzen 7 PRO 7840HS w/ Radeon 780M Graphics 
                   │     base     │                new                 │
                   │    sec/op    │    sec/op     vs base              │
LogEvents/sample=0   1.477µ ± 23%   1.432µ ± 24%       ~ (p=0.621 n=6)
LogEvents/sample=1   1.483µ ± 78%   1.451µ ± 90%       ~ (p=0.240 n=6)
LogEvents/sample=2   2.372µ ±  5%   2.331µ ±  1%  -1.75% (p=0.006 n=6)
geomean              1.732µ         1.691µ        -2.33%

                   │      base      │                  new                  │
                   │      B/op      │      B/op       vs base               │
LogEvents/sample=0   1.532Ki ±  97%   1.343Ki ± 112%        ~ (p=0.457 n=6)
LogEvents/sample=1    1064.0 ± 454%     784.0 ± 651%  -26.32% (p=0.013 n=6)
LogEvents/sample=2   1.551Ki ±   0%   1.059Ki ±   0%  -31.74% (p=0.002 n=6)
geomean              1.352Ki          1.029Ki         -23.90%

                   │    base    │               new                │
                   │ allocs/op  │ allocs/op   vs base              │
LogEvents/sample=0   11.00 ± 0%   10.00 ± 0%  -9.09% (p=0.002 n=6)
LogEvents/sample=1   11.00 ± 0%   10.00 ± 0%  -9.09% (p=0.002 n=6)
LogEvents/sample=2   13.00 ± 0%   12.00 ± 0%  -7.69% (p=0.002 n=6)
geomean              11.63        10.63       -8.63%
```

#### Which issue(s) this PR fixes:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #

#### Special notes for your reviewer:

#### Does this PR introduce a user-facing change?
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".

For more information on release notes see: https://git.k8s.io/community/contributors/guide/release-notes.md
-->
```release-note
NONE
```

#### Additional documentation e.g., KEPs (Kubernetes Enhancement Proposals), usage docs, etc.:

<!--
This section can be blank if this pull request does not require a release note.

When adding links which point to resources within git repositories, like
KEPs or supporting documentation, please reference a specific commit and avoid
linking directly to the master branch. This ensures that links reference a
specific point in time, rather than a document that may change over time.

See here for guidance on getting permanent links to files: https://help.github.com/en/articles/getting-permanent-links-to-files

Please use the following format for linking documentation:
- [KEP]: <link>
- [Usage]: <link>
- [Other doc]: <link>
-->
```docs

```