@vrutkovs 2025-05-08
/remove-lifecycle rotten